+++ src/components/Dashboard/profile/trust-engine/CommunityRanking.tsx
@@ -1,10 +1,10 @@
 "use client"
 
 import { ArrowDown, ArrowUp } from "lucide-react"
-import { Card } from "../ui/card"
-import { Avatar, AvatarFallback } from "../ui/avatar"
-import { Badge } from "../ui/badge"
-import { Progress } from "../ui/progress"
+import { Card } from "../../../ui/card"
+import { Avatar, AvatarFallback } from "../../../ui/avatar"
+import { Badge } from "../../../ui/badge"
+import { Progress } from "../../../ui/progress"
 import { CommunityRankingsData } from "./Constant"
 
 export function CommunityRanking() {
